+++ src/Pagination/Pagination.tsx
@@ -84,7 +84,7 @@ export const getMuiTablePaginationThemeOverrides = (): Components<Omit<Theme, 'c
                     display: 'none', // hides the form action label space from Autocomplete
                   },
                   '> .autocomplete-container': {
-                    margin: '0 12px 0 4px',
+                    margin: '0 4px',
                   },
                 },
                 'div[data-testid=tablePaginationActionsPageDiv]': {