+++ src/List/ListItemText.tsx
@@ -37,6 +37,11 @@ export const getMuiListItemTextThemeOverrides = (): Components<Omit<Theme, 'comp
         root: ({ ownerState, theme }) => {
           return {
             margin: '0',
+            '&.list-item-text-truncate .MuiTypography-root': {
+              overflow: 'hidden',
+              textOverflow: 'ellipsis',
+              whiteSpace: 'nowrap',
+            },
           };
         },
       },
@@ -98,6 +103,7 @@ const ListItemText = ({
 
   return (
     <MuiListItemText
+      className={`${tooltip ? '' : 'list-item-text-truncate'}`}
       ref={listRef}
       {...props}
       primary={primaryReactNode}