+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 ### Changed
+- Added keyboard accessibility to the header in `DataGrid` component
 
 ### Breaking changes
 
@@ -53,7 +54,6 @@
 - Added `tooltipPlacement` prop to the `MultipleSelectChip`, `Autocomplete` and `Panel` component. 
 - Updated icon button spacing in Snackbar
 - Adjusted checkbox alignment in `DataGrid` component
-- Added keyboard accessibility to the header in `DataGrid` component
 
 ### Breaking changes