+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 - Added keyboard accessibility to the header in `DataGrid` component
 - Added focus to the input label in `TextField` and `SelectMultiple` components
+- Fixed the width of the `Tile` action menu and corrected the focus of nested-level `Accordion`
 
 ### Breaking changes