+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ### Fixed
 
 ### Changed
+- Added keyboard accessibility to the header in `DataGrid` component
+- Added focus to the input label in `TextField` and `SelectMultiple` components
+- Fixed the width of the `Tile` action menu and corrected the focus of nested-level `Accordion` component
 
 ### Breaking changes