+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - Added `tooltipPlacement` prop to the `MultipleSelectChip`, `Autocomplete` and `Panel` component. 
 - Updated icon button spacing in Snackbar
 - Adjusted checkbox alignment in `DataGrid` component
+- Resolved tooltip issue of `Autocomplete` component
 
 ### Breaking changes